Extract User model routes
