+++ routes/user/user.js
@@ -1,0 +1,16 @@
+const express = require("express")
+const router = express.Router()
+const auth = require("../../middleware/auth");
+const User = require("../../db/models/user")
+
+router.get("/", auth, async (req, res) => {
+    try {
+      const user = await User.findById(req.user.id).select("-password");
+      res.json(user);
+    } catch (err) {
+      console.error(err.message);
+      res.status(500).send("Server error");
+    }
+  });
+
+module.exports = router
